Add 2019 API support, refactor

+++ app/src/main/java/pl/temomuko/autostoprace/ui/base/drawer/DrawerMvpView.java
@@ -11,5 +11,5 @@ public interface DrawerMvpView extends MvpView {
 
     void setupHeaderEmail(String email);
 
-    void setupTeamNumberText(int teamNumber);
+    void setupTeamNumberText(long teamNumber);
 }
